backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.6.1
SQLAlchemy==2.0.36
asyncpg==0.30.0
alembic==1.14.0
psycopg[binary]==3.2.3
paho-mqtt==2.1.0
httpx==0.27.2
networkx==3.4.2

